pyproject.toml
[project]
name = "llm-batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chevron>=0.14.0",
    "cyclopts>=3.14.0",
    "openai>=1.76.0",
    "polars>=1.28.1",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
]

[project.scripts]
openai-batch = "openai_batch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openai_batch"]
